Updating parsing
Adding auxiliary modules for supporting a real database. It includes the top module, the flat module, and the flat module with parameters explicitly included, so the module is executable without requiring the Maude database. Most of the functions need to be implemented, but the data structures for storing it are available.

State included in the I/O object, so we can control the flow.